Update __init__.py (Indexer)
All previous versions included a mistake that was not made clear until testing in a multi-indexer environment; the elasticsearch repository create operations expect exclusivity in controlling the GCS directory thus expecting the directory to be in the last state they left it in, any tampering of this state immediately strikes an repository_exception. This is now handled by dynamically adapting the local repository whenever this type of exception appears.